+++ example/lib/pages/inline_floats.dart
@@ -2,7 +2,7 @@ import 'package:float_column/float_column.dart';
 import 'package:flutter/material.dart';
 
 class InlineFloats extends StatelessWidget {
-  const InlineFloats({Key? key}) : super(key: key);
+  const InlineFloats({super.key});
 
   @override
   Widget build(BuildContext context) {