+++ lib/src/shared.dart
@@ -3,4 +3,5 @@
 // LICENSE file.
 
 enum FCFloat { none, left, right, start, end }
+
 enum FCClear { none, left, right, start, end, both }